requirements.txt
# Yahoo Fantasy Baseball CLI dependencies
yahoo_fantasy_api>=2.12.0
yahoo-oauth>=2.0

# Data analysis
pandas>=2.1.0
numpy>=1.26.0

# Baseball data
pybaseball>=2.2.7
MLB-StatsAPI>=1.7.0

# Browser automation (write operations fallback)
playwright>=1.40.0

# API server
flask>=3.0.0

# Automation config
PyYAML>=6.0
